+++ .gitignore
@@ -5,3 +5,4 @@ node_modules
 .next
 build/
 *.off
+.vscode
