+++ package.json
@@ -20,17 +20,17 @@
     "@testing-library/react": "^11.1.0",
     "@testing-library/user-event": "^12.1.10",
     "axios": "^0.27.2",
+    "eslint-config-next": "^13.2.3",
     "eslint-config-react-app": "^7.0.1",
     "hamburger-react": "^2.5.0",
     "moment": "^2.29.4",
-    "next": "^12.3.1",
-    "react": "^17.0.2",
+    "next": "^13.2.3",
+    "react": "^18.2.0",
     "react-circular-progressbar": "^2.1.0",
-    "react-dom": "^17.0.2",
+    "react-dom": "^18.2.0",
     "react-loading-icons": "^1.1.0",
     "react-moment": "^1.1.2",
     "react-parallax": "^3.5.1",
-    "react-swipeable-views": "^0.14.0",
     "react-table": "^7.7.0",
     "typewriter-effect": "^2.19.0",
     "web-vitals": "^2.1.4"
@@ -41,7 +41,7 @@
     "start": "next start"
   },
   "engines": {
-    "node": "14.x"
+    "node": "16.x"
   },
   "eslintConfig": {
     "extends": [
