+++ src/components/HackathonList/EventFeature.js
@@ -25,12 +25,15 @@ function EventFeature(props) {
     end_date,
     location,
     devpostUrl,
-    eventLinks,
+    rawEventLinks,
     donationUrl,
     donationGoals,
     donationCurrent,
     icon,
   } = props;
+
+  // TODO: Is the schema on the backend wrong? Or is the schema here wrong?
+  const eventLinks = typeof rawEventLinks === 'string' ? [rawEventLinks] : rawEventLinks
 
   return (
     <EventCards container direction="column">
