+++ src/pages/admin/index.js
@@ -1,41 +1,11 @@
-import React from "react";
+import dynamic from 'next/dynamic';
 
-import AddNonProfit from "../../components/admin/nonprofit-add";
-import AddProblemStatement from "../../components/admin/problemstatement-add";
-import AddHackathon from "../../components/admin/hackathon-add";
-import AdminHackathonList from "../../components/admin/hackathon-list";
+const Admin = dynamic(() => import('../../components/admin/Admin'), {
+    ssr:false,
+});
 
-import useAdmin from '../../hooks/use-admin-check'
-import useNonprofit from "../../hooks/use-nonprofit";
-import useTeams from "../../hooks/use-teams";
-import useProblemstatements from "../../hooks/use-problem-statements";
-import useHackathonEvents from "../../hooks/use-hackathon-events";
-
-import Head from 'next/head';
-
-export default function Admin ({ admin }) {    
-    const { isAdmin } = useAdmin();
-    const { nonprofits } = useNonprofit();
-    const { teams } = useTeams();
-    const { problem_statements } = useProblemstatements();
-    const { hackathons } = useHackathonEvents();
-
-    if (!isAdmin)
-    {
-        return(<span></span>);
-    } 
-
-    console.log("admin.js repaint");
-
+export default function AdminPage ({ admin }) {    
     return(
-        <div>
-        <Head>
-            <title>Opportunity Hack Developer Portal</title>
-        </Head>
-            <AdminHackathonList hackathons={hackathons} problem_statements={problem_statements} />
-            <AddProblemStatement />            
-            <AddHackathon nonprofits={nonprofits} teams={teams} />             
-            <AddNonProfit problem_statements={problem_statements} />            
-        </div>
-        );
+        <Admin/>      
+    );
 };
